+++ ruitk/themes/fallen/base.js
@@ -29,24 +29,24 @@ let config = {
 
     font : "Arial, Helvetica, sans-serif",
        
-    fontSizeH1 : "3.2vw",
-    fontSizeH2 : "2.5vw",
-    fontSizeH3 : "2vw",
-    fontSizeP1 : "1.5vw",
-    fontSizeP2 : "1vw",
-    fontSizeP3 : "0.75vw",
+    fontSizeH1 : 3.2,
+    fontSizeH2 : 2.5,
+    fontSizeH3 : 2,
+    fontSizeP1 : 1.5,
+    fontSizeP2 : 1,
+    fontSizeP3 : 0.75,
 
-    borderRadius : "1vw",
-    borderWidth : "0.25vw",
+    borderRadius : 1,
+    borderWidth : 0.25,
 
-    marginSmall : "0.1vw",
-    paddingSmall : "0.1vw",
+    marginSmall : 0.1,
+    paddingSmall : 0.1,
 
-    marginMedium : "0.6vw",
-    paddingMedium : "0.6vw",
+    marginMedium : 0.6,
+    paddingMedium : 0.6,
 
-    marginLarge : "1vw",
-    paddingLarge : "1vw",
+    marginLarge : 1,
+    paddingLarge : 1,
 };
 
 let elements = [
@@ -62,13 +62,15 @@ let elements = [
                 h : "auto",
                 c : "",
                 r : "",
-                xAline : "",
-                yAline : "",
+                xAlign : "",
+                yAlign : "",
                 textAlign : "left",
                 position : "",
+                element : element.element
             }, info);
         
-            let e = document.createElement(element.element);
+            let e = document.createElement(info.element);
+            delete info.element;
             e.id = info.id;
             
             e.style.textAlign = info.textAlign;
@@ -88,7 +90,7 @@ let elements = [
             }
 
 
-            if (info.x === "" && info.y === "" && info.xAline === "" && info.yAline === "" && info.w === "auto" && info.h === "auto" ) {// all defaults
+            if (info.x === "" && info.y === "" && info.xAlign === "" && info.yAlign === "" && info.w === "auto" && info.h === "auto" ) {// all defaults
                 e.style.position = "relative";
             } else {
                 e.style.position = "absolute";
@@ -113,14 +115,14 @@ let elements = [
                 bottom: 100
             };
             // Handle horizontal alignment
-            if (info.xAline !== "") {
-                let xPercent = convert[info.xAline];
+            if (info.xAlign !== "") {
+                let xPercent = convert[info.xAlign];
                 e.style.left = `${xPercent}%`;
                 e.style.transform += `translateX(-${xPercent}%)`;
             }
             // Handle vertical alignment
-            if (info.yAline !== "") {
-                let yPercent = convert[info.yAline];
+            if (info.yAlign !== "") {
+                let yPercent = convert[info.yAlign];
                 e.style.top = `${yPercent}%`;
                 e.style.transform += ` translateY(-${yPercent}%)`;
             }
@@ -131,7 +133,7 @@ let elements = [
             transition: "var(--transition)",
             fontFamily : "var(--font)",
             focus_outline : "3px solid var(--standout3)",
-            fontSize : "1vw",
+            fontSize : "var(--fontSizeP2)",
             position : "relative",
             overflow : "hidden",
             boxSizing : "border-box",
@@ -319,7 +321,19 @@ let elements = [
 
 function init () {
     Style.declare(colors);
-    Style.declare(config);
+    let zoomer = () => {
+        let hold = {};
+        for (let key in config) {
+            if (typeof config[key] === "number") {
+                hold[key] = `${config[key] * window.devicePixelRatio}vw`;
+                continue;
+            };
+            hold[key] = config[key];
+        };
+        Style.declare(hold);
+    };
+    zoomer();
+    window.addEventListener('resize', zoomer);
 }
 
 export class FallenBase {