+++ doc/changelog.md
@@ -1,6 +1,30 @@
+# 0.9
+#### Fallen Changes
+ - Renamed `xAline` and `yAline` to `xAlign` and `yAlign` to be more constant with `textAlign`
+ - Made it so zoom actually zooms
+ - Completed the following elements in staging
+	 - `radio`
+	 - `checkbox`
+	 - `dropdown`
+	 - `combo`
+	 - `button`
+	 - `submit`
+#### support changes
+ - fixed some bugs in `convert.js` that was causing outputted strings to be incorrectly when converting to `camelCase` 
+ - Added some input filtering to `convert.js`
+ - Added a new dependency called `tester.js` for input validation
+#### element.js Changes
+ - Added error handling to for failed elements previously 1 failure would cause everything to fail to render
+ - Added support for functions and arrow functions
+ - Added some ASCII art to init
+ - Added input filtering for the parse function
+#### documentation Changes
+ - Completed documentation the following fallen elements
+	 - `radio`
+	 - `checkbox`
 ## 0.8.5
 #### documentation Changes
--  made documentation for [convert.js](support/convert.js.md)
+- made documentation for [convert.js](support/convert.js.md)
 - reorganized the documentation for [fallen](fallen/Fallen%20Summary.md)
 ## 0.8.4
 #### support changes